# docs/swagger.yaml
basePath: /
definitions:
  service.CreateUserInput:
    properties:
      password:
        type: string
      role:
        type: string
      username:
        type: string
    required:
    - password
    - role
    - username
    type: object
  service.CreateVoucherInput:
    properties:
      code:
        type: string
      discount_amount:
        type: number
      quota:
        type: integer
    required:
    - code
    - discount_amount
    - quota
    type: object
  service.LoginInput:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  service.TicketVariantInput:
    properties:
      end_date:
        type: string
      name:
        type: string
      price:
        minimum: 0
        type: number
      quota:
        minimum: 1
        type: integer
      start_date:
        type: string
    required:
    - name
    - price
    - quota
    type: object
  service.UpdateVoucherInput:
    properties:
      discount_amount:
        type: number
      quota:
        type: integer
    required:
    - discount_amount
    - quota
    type: object
  utils.Meta:
    properties:
      code:
        type: integer
      message:
        type: string
      status:
        type: string
    type: object
  utils.SuccessResponse:
    properties:
      data: {}
      meta:
        $ref: '#/definitions/utils.Meta'
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API Documentation for SMILE FEST 2026 Ticketing System.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: SMILE FEST API
  version: "1.0"
paths:
  /api/admin/ticket-variants:
    get:
      description: Mengambil seluruh data tipe tiket tanpa peduli status aktif atau
        periode tanggal.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.SuccessResponse'
      security:
      - BearerAuth: []
      summary: Get All Ticket Variants (Admin)
      tags:
      - admin
    post:
      consumes:
      - application/json
      description: Admin membuat gelombang tiket baru
      parameters:
      - description: Data Gelombang Tiket
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/service.TicketVariantInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/utils.SuccessResponse'
      security:
      - BearerAuth: []
      summary: Create New Ticket Variant
      tags:
      - admin
  /api/admin/ticket-variants/{id}:
    delete:
      description: Admin menghapus gelombang tiket (Hanya jika belum ada transaksi
        terkait)
      parameters:
      - description: Variant ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.SuccessResponse'
      security:
      - BearerAuth: []
      summary: Delete Ticket Variant
      tags:
      - admin
    put:
      consumes:
      - application/json
      description: Admin mengubah harga atau periode gelombang tiket
      parameters:
      - description: Variant ID
        in: path
        name: id
        required: true
        type: string
      - description: Data Update Gelombang Tiket
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/service.TicketVariantInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.SuccessResponse'
      security:
      - BearerAuth: []
      summary: Update Ticket Variant
      tags:
      - admin
  /api/admin/ticket-variants/{id}/toggle:
    put:
      description: Admin membuka atau menutup penjualan fase tiket secara manual (Kill
        switch)
      parameters:
      - description: Variant ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.SuccessResponse'
      security:
      - BearerAuth: []
      summary: Toggle Ticket Variant Status
      tags:
      - admin
  /api/admin/users:
    get:
      description: Admin melihat daftar seluruh panitia dan admin yang terdaftar
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.SuccessResponse'
      security:
      - BearerAuth: []
      summary: Get All Users
      tags:
      - admin
    post:
      consumes:
      - application/json
      description: Hanya bisa diakses Admin untuk membuat akun Scanner/Admin baru
      parameters:
      - description: Data User Baru
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/service.CreateUserInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.SuccessResponse'
      security:
      - BearerAuth: []
      summary: Create New User
      tags:
      - admin
  /api/admin/users/{id}:
    delete:
      description: Admin menghapus/mencabut akses akun panitia
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.SuccessResponse'
      security:
      - BearerAuth: []
      summary: Delete User
      tags:
      - admin
  /api/admin/vouchers:
    get:
      description: Admin melihat daftar voucher dan sisa kuota
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.SuccessResponse'
      security:
      - BearerAuth: []
      summary: Get All Vouchers
      tags:
      - admin
    post:
      consumes:
      - application/json
      description: 'Admin menambah kode diskon baru (misal: PELAJAR20K)'
      parameters:
      - description: Data Voucher
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/service.CreateVoucherInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.SuccessResponse'
      security:
      - BearerAuth: []
      summary: Create New Voucher
      tags:
      - admin
  /api/admin/vouchers/{id}:
    delete:
      description: Admin menghapus voucher secara permanen
      parameters:
      - description: Voucher ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.SuccessResponse'
      security:
      - BearerAuth: []
      summary: Delete Voucher
      tags:
      - admin
    put:
      consumes:
      - application/json
      description: Admin merubah potongan harga atau kuota voucher
      parameters:
      - description: Voucher ID
        in: path
        name: id
        required: true
        type: string
      - description: Data Update Voucher
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/service.UpdateVoucherInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.SuccessResponse'
      security:
      - BearerAuth: []
      summary: Update Voucher
      tags:
      - admin
  /api/admin/vouchers/{id}/toggle:
    put:
      description: Admin mematikan (Kill Switch) atau menyalakan kembali voucher
      parameters:
      - description: Voucher ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.SuccessResponse'
      security:
      - BearerAuth: []
      summary: Toggle Voucher Status
      tags:
      - admin
  /api/login:
    post:
      consumes:
      - application/json
      description: Mendapatkan JWT token untuk akses API
      parameters:
      - description: Kredensial Login
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/service.LoginInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.SuccessResponse'
      summary: User Login
      tags:
      - auth
  /api/scanner/stats:
    get:
      description: Panitia lapangan melihat jumlah peserta yang sudah masuk
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.SuccessResponse'
      security:
      - BearerAuth: []
      summary: Get Scanner Stats
      tags:
      - scanner
  /api/scanner/validate-ticket:
    post:
      consumes:
      - application/json
      description: Memvalidasi QR Code dari tiket peserta di hari H
      parameters:
      - description: Ticket UUID
        in: body
        name: input
        required: true
        schema:
          additionalProperties:
            type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.SuccessResponse'
      security:
      - BearerAuth: []
      summary: Validate Ticket
      tags:
      - scanner
  /api/seed-admin:
    post:
      description: Temporer untuk membuat akun admin pertama (Matikan rute ini setelah
        dipakai)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.SuccessResponse'
      summary: Seed First Admin
      tags:
      - auth
  /api/tickets/info:
    get:
      description: Mengambil data harga dan tipe tiket yang sedang aktif untuk publik.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.SuccessResponse'
      summary: Get Ticket Info
      tags:
      - public
  /api/tickets/track:
    get:
      description: Peserta mencari tiket mereka, mengecek status verifikasi admin,
        atau melihat instruksi pembayaran manual.
      parameters:
      - description: Order ID (SMILE-xxx)
        in: query
        name: order_id
        required: true
        type: string
      - description: Email Peserta
        in: query
        name: email
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.SuccessResponse'
      summary: Track E-Ticket & Payment Status
      tags:
      - public
  /api/vouchers/validate:
    get:
      description: Peserta mengecek apakah voucher valid dan kuota masih ada sebelum
        checkout
      parameters:
      - description: Kode Voucher
        in: query
        name: code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.SuccessResponse'
      summary: Validate Voucher Code
      tags:
      - public
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"